ai-service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.4
pydantic_core==2.33.2
langchain==0.3.25
langchain-core==0.3.59
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
openai==1.79.0
python-dotenv==1.0.0
pyyaml==6.0.1
boto3==1.38.23
aiofiles==24.1.0